fix:修复一些重构引起的问题

+++ mooc-authorize2/src/main/java/cn/edu/lingnan/mooc/authorize/model/param/ManagerParam.java
@@ -16,7 +16,10 @@ import java.util.List;
 @ToString
 public class ManagerParam {
 
-    private Integer id;
+    /**
+     * 管理员id
+     */
+    private Long id;
     // 名字
     private String name;
     // 登录账号
